fix(remix): Export `Integration` type declaration as union type (#8016)
Change the isomorphic type declarations of `Sentry.Integrations` to a union type rather than an object. This stops `tsc` from resolving the types of the individual integration types which caused errors due to incorrect path resolvings of dependencies of `@sentry/remix`. With this change we now declare `Sentry.Integrations` in Remix identically to how we do it in NextJS and SvelteKit.