fix(replay): Ignore max session life for buffered sessions
Theres an edge case where a buffered session becomes expired, an error comes in, the link from error to replay is lost because a new session is created due to the session being expired.

We should either ignore the max session life for a buffered session, or possibly check/refresh session when an error comes in.