feat(sveltekit): Add source maps support for Vercel (lambda) (#8256)
Adjust our automatic source maps upload setup in the SvelteKit
SDK to support SvelteKit apps deployed to Vercel. This will only work
for Lambda functions/Node runtime; not for the Vercel Edge runtime.

This required a few changes in our custom vite plugin as well as on the
server side of the SDK:

* Based on the used adapter (manually set or detected via #8193) and the
`svelte.config.js` we determine the output directory where the generated
JS emitted to.
* The determined output directory is injected into the global object on
the server side
* When an error occurs on the server side, we strip the absolute
filename of each stack frame so that the relative path of the
server-side code within the output directory is left.
* We also use the determined output directory to build the correct
`include` entries for the source map upload plugin.

With this change, source maps upload should work for auto and Vercel
adapters, as well as for the Node adapter.
As for the Node adapter, the stackframe rewrite behaviour was also
changed but it is now more in line with all supported adapters.