fix(remix): Don't log missing parameters warning on server-side. (#8269)
We are logging a warning when any of the required hooks / functions to
Remix Browser SDK are not available in the Root component.

This Root component is also rendered on server-side where we don't have
access to those functions. And as we're not creating pageload /
navigation transactions there, those functions are not required anyway.

So this PR prevents that warning to be logged on server-side which is
misleading.