feat(replay): Do not restart replays on "keydown" (#8226)
This is causing some low-value replays. If you search sentry for replays
with duration < 5 seconds, you will see that most of them are re-newed
sessions from idle that only have a "keydown" breadcrumb. I suspect
users are alt-tabbing out of an idle Sentry page. Removing for now until
we have a better way to detect when a session is truly resumed via
keydown event.