feat(tracing): Add experiment to capture http timings
This adds an experiment that will allow http timings to be captured. We currently capture timings on Sentry SaaS with some custom code and append them to the spans, which has been helpful to identify some performance problems that were previously hidden (http/1.1 stall time). We will likely change the span waterfall to represent some of these as subtimings and are also considering adding a http/1.1 stall performance issue.