feat(replay): Rework slow click & multi click detection (#8322)
This PR reworks the slow click detection to accommodate rage click
detection as well.
This required substantial changes, as we need to keep track of stuff
much more.

Now, we keep a list of clicks that come in in a new class. We register a
single set of listeners (mutation observer, click listener, scroll
listener), and then try to route things to the correct clicks as much as
possible.

Any clicks within 1 second count as "multi click", so are not considered
for slow clicks at all, but counted on the first click. After a second,
a click (even on the same element) will be treated as a "new" click.

ref https://github.com/getsentry/sentry-javascript/issues/8300