feat(eslint): Add `no-return-await` rule to eslint config (#8388)
Add https://eslint.org/docs/latest/rules/no-return-await to eslint
config to remove usage of uneeded async/await calls. This helps reduce
microtasks being generated, which can help reduce memory pressure caused
by the SDK.

The downside of removing `return await` is that stacktraces get slightly
worse for async errors that use these methods, as we no longer pause
execution on return for the engine to grab context on, but instead just
pass through the promise, but I think it's worth it for this to be the
default, and for us to opt-in to the better stacktraces if need be.