build: Update typescript from 3.8.3 to 4.9.5 (#8255)
This updates the TS version we use from 3.8.3 to 4.9.5. This allows us
to use newer TS features, as well as as improves compatibility with
newer @types/xxx packages that use newer features, which leads to issues
when transitive dependencies are used etc.

We use [downlevel-dts](https://www.npmjs.com/package/downlevel-dts) to
generate a 3.8 compatible types output _in addition_, so this should be
fully backwards compatible. TS provides a built-in way to tell it to use
a certain types output for a certain version, so this should work with
no breaking change.

The main changes necessary here are small tweaks to types which are
required in 4.9.x.