feat(node): Populate propagation context using env variables (#8422)
Requires https://github.com/getsentry/sentry-javascript/pull/8421 to be
merged

ref https://github.com/getsentry/sentry-javascript/issues/8352

This PR adds support for
https://github.com/getsentry/rfcs/blob/main/text/0071-continue-trace-over-process-boundaries.md
via propagation context. When the Node SDK initializes, it grabs
`process.env.SENTRY_TRACE` and `process.env.SENTRY_BAGGAGE` and uses
them to populate the existing propagation context.

In future PRs (tracked by #8352), we will be adding support for
transactions/outgoing requests to use propagation context, but for now
we are only storing in propagation context, not doing anything with it.