build(angular): Avoid loading all `@types/` packages
As `@sentry/angular` uses TS 4.0, but the default behavior of TS is to load _all_ `@types/xxx` packages it can find in node_modules, the build will fail if _any_ other package uses a types package that uses newer TS features. 

In order to avoid this, we opt out of the automatic types resolution - since we are not using any `@types/xxx` package in angular as of now, this doesn't change anything.