feat(replay): Ensure min/max duration when flushing (#8596)
This PR adds a safeguard to ensure we do not flush (=send) a replay that
is either too short or too long.

We allow to configure a `minReplayDuration`, which defaults to 5s and
maxes out at 15s. Whenever we try to flush and the duration is shorter
than this, we'll just skip flushing.

Additionally, we also skip flushing if the replay is longer than
MAX_SESSION_LIFE + 5s (=60min + 5s). This _should not_ happen,
technically, but apparently it still does. So while we figure out the
root cause of this, we can at least avoid sending stuff in that case.