test(e2e): Use `latest || *` instead of `*` as version (#8585)
Due to the [way pnpm handles
versions](https://github.com/pnpm/pnpm/issues/6463), the version `*` we
use in E2E tests is actually incorrect, because it installs the _lowest_
version it can find, e.g. 0.1.0. This is usually not a problem as when
we use verdaccio, there is only a single version in the repository.
however, when running things locally/debugging stuff, and you run `pnpm
install` without verdaccio, stuff fails. This updates the versions used
in E2E tests to `latest || *`, which results in a correct resolution.