fix(utils): Truncate aggregate exception values (LinkedErrors) (#8593)
Aggregate exceptions' `exception.value` strings
weren't truncated correctly in our event preparation pipeline. The
tricky part here is that the `LinkedErrors` integration adds the
linked/child exceptions to the event in an event processor which is
applied only after we truncate the original event. Hence, we need to
also truncate the values in the event processor itself.