feat(node-experimental): Add `@sentry/node-experimental` package as MVP for POTEL (#8609)
This introduces a new package `@sentry/node-experimental`, which is an
MVP implementation for Performance powered by OTEL (POTEL).

This package provides a wrapper over `@sentry/node` (also using
`@sentry/opentelemetry-node`) which uses opentelemetry for performance
instrumentation under the hood.

This is all abstracted away from the user, they only need to do:

```js
import * as Sentry from '@sentry/node-experimental';

Sentry.init({
  dsn: 'my-dsn',
  tracesSampleRate: 0.1
});
```

And it will set up all the necessary integrations etc. for the user,
including the default integrations + any applying performance
integrations.

For this, I added all performance integrations that exist for node,
minus undici (which will take some more work to get going), plus some
new stuff: mysql, fastify, nest. Also, mongoose is it's own integration
now, not handled via mongo anymore.

Note that I've only manually tested express & http so far. All others
still need to be verified etc. But this is really a POC, which may go
somewhere or not - by publishing this, it becomes super easy to try this
in various scenarios.

I have tweaked the Http integration to create more or less the same
output as we're used to for Sentry. I also made small tweaks to
`@sentry/opentelemetry-node` to allow us to communicate with this a bit
better. There are two main use cases I've identified there:

* Tell the span processor to drop a span (=not send it to Sentry). For
this, I added dedicated hooks.
* Add some specific data/context/tags/metadata to an otel span, that
should be added to the resulting sentry span. For this, I added utils
which keep a map of data to be added in `spanEnd`.

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>