fix(node-experimental): Update auto integration lookup & readme (#8690)
This updates the way we auto-lookup integrations.

Due to how OTEL instrumentation works, we actually do not need any
dynamic require. Otel instrumentation will generally always work, it
only actually instruments when something is required/imported. So all
integration _will_ be added, they will simply do nothing if the package
is not used.

This also updates the readme to clarify this & restrictions.