chore(ci): Make "Required Tests passed" job depend on artifact upload and other required jobs (#8751)
Adjust our final "Required Tests Passed" CI job to not only
depend on required _test_ jobs but on _all_ required jobs. Specifically,
this adds

* Lint
* Circular deps check
* Most importantly: Upload Artifacts

With this change, we can configure craft's status provider to
specifically wait for this job which should fix the artifacts download
timeout (https://github.com/getsentry/craft/issues/482). It's worth
noting that this PR will only fix this timeout in our repo and in a way
we're adjusting to the status provider check. However, given that we're
apparently the only repo where this happens, it's probably justified and
it makes things more explicit.