fix(replay): Ensure we do not try to flush when we force stop replay (#8783)
In our `stop()` method, we always tried to force a final flush when in
`session` mode.
However, that may lead to weird behaviour when we internally `stop()`
due to a failure - e.g. think a send replay request fails, we do not
want to force a flush again in that case.

Note that in tests this seems to be generally passing because `flush()`
usually has a running `_flushLock` at this time and thus does not
attempt to flush again immediately but only schedules a flush later. I
added a test that properly failed for this before and is now fixed.