fix(nextjs): Ensure imports are valid relative paths
Fixes https://github.com/getsentry/sentry-javascript/issues/8798

In our logic that injects the Sentry config files to be executed, `path.relative()` may return something like `sentry.server.config.js` which is not allowed. Imports from the current directory need to start with './'. This is why we prepend the path with './', which should always again be a valid relative path.