fix(replay): Streamline session creation/refresh (#8813)
We've been using `_loadAndCheckSession` both in initial session setup as
well as when checking for expiration of session.
This leads to some not-so-optimized stuff, as we kind of have to do
double duty in there (e.g. we constantly re-assign the session etc).

This streamlines this by splitting this into:

* `_initializeSessionForSampling()`: Only called in
`initializeSampling()`
* `_checkSession()`: Called everywhere else, assumes we have a session
setup yet

Only the former actually looks into sessionStorage, the latter can
assume we always have a session already.

This also extends the behavior so that if we fetch a `buffer` session
from storage and segment_id > 0, we start the session in `session` mode.
Without this, we could theoretically run into endless sessions if the
user keeps refreshing and keeps having errors, leading to continuous
switchovers from buffer>session mode.