fix(node): Log entire error object in `OnUncaughtException` (#8876)
In our `OnUncaughtException` integration, we have to emulate Node's
default behaviour of logging errors to the console.
For some reason though, we only logged the stack trace of an `error`
with a stack trace instead of the entire error object.

As reported in #8856, this causes additional properties on the error
(such as `cause`) not to be logged anymore which doesn't reflect Node's
default behaviour (see issue for comparison). This patch simplifies the
`console.error` call to just always log the entire `error`.