ref(tracing-internal): Deprecate `tracePropagationTargets` in `BrowserTracing` (#8874)
This PR deprecates `BrowserTracing`'s `tracePropagationTargets` option
in favour of the top-level Sentry.init option.

When introducing Tracing without Performance, we opted to [promote
`tracePropagationTargets` to a top-level
option](https://github.com/getsentry/sentry-javascript/pull/8395). We
deprecated the integration-level option in Node's `Http` integration but
not in `BrowserTracing`. This patch fixes that.