fix(browser): Add replay and profiling options to `BrowserClientOptions` (#8921)
Add missing replay and profiling options to the
`BrowserClientOptions` interface. Previously, we only exposed these
types to `BrowserOptions`. This led to type errors for users who
directly create a client, without using `Sentry.init` (as reported in
#8857).

While one would think that `BrowserClientOptions` is basically a
superset of `BrowserOptions` (leaving aside `stackparser`,
`integrations` and `transport`) this is in fact not the case, which IMO
is the core problem here. `BrowserClientOptions` only inherits the base
`Options` (which are shared between browser and node), in addition to
transport options. However, changing this so that,
`BrowserClientOptions` inherits from `BrowserOptions` is a breaking
change, so I opted to just add the missing options to
`BrowserClientOptions`.