ref(node-otel): Avoid exporting internals & refactor attribute adding (#8920)
I noticed that we kind of extended the public API by exporting some
things from opentelemetry-node that we need in node-experimental. I am
deprecating these internals as I'm refactoring some things to be a bit
cleaner IMHO.

The one export we actually need outside I am exporting as
`_INTERNAL_getSentrySpan` to make it clear this is not meant to be used
otherwise. The main reason I am not just exporting this is that this
_may_ have to change if we change how we internally store/handle spans -
e.g. if we move away from the map this is just not possible to have
anymore, and I do not want to lock us into having to support this in the
future.