feat(node-experimental): Implement new performance APIs (#8911)
This re-implements the new performance APIs:

* `startActiveSpan`
* `startSpan`
* `getActiveSpan`

In `@sentry/node-experimental`, where these delegate to OTEL under the
hood.
I tried to make sure to have _exactly_ the same APIs as we have for
these functions without this. There are some caveats with this:

* calling `span.finish()` may behave in unexpected ways - it is not
supported to manually finish a span.
* Calling `scope.getSpan()` will not work
* Calling `span.startSpan()` will not work

Basically, only the top level exported methods work, nothing else.