feat(node-experimental): Sync OTEL context with Sentry AsyncContext (#8797)
This PR implements a strategy to sync the OpenTelemetry Context with our
own Hub forking for AsyncContext.

This works by fully relying on OpenTelemetry to handle async context
isolation/forking.

1. We implement a custom OTEL ContextManager that wraps the default
AsyncHooks manager, but makes sure to also fork the hub for each context
change & puts the hub on the OTEL context so we can retrieve it.
2. Then, we also have a custom Sentry AsyncContextStrategy which just
refers to OTEL context and picks the hub from there we put there in 1.

This means we do not need to do any context forking ourselves anymore,
so no need for e.g. `Sentry.Handlers.requestHandler()` and stuff like
this.

It _should_ also mean that Sentry & OTEL should be as in sync as
possible.

Some notes:
* Currently only works for AsyncHooks, which should be fine I guess.
Could also be exteded to work with other implementations as well.