fix(node-experimental): Ensure http breadcrumbs are correctly added
Due to the way OTEL handles context (=scope/hub), I noticed that breadcrumbs added in the http integration are not correctly captured.
The reason is that OTEL forks the hub for the http request context, and runs the hooks inside of this forked context.
In order to deal with this, we have to actually attach the breadcrumb to the _parent_ Hub in this specific case.