fix(browser): Check for existence of instrumentation targets (#8939)
There are cases when global objects such as the window object are
shimmed that they define properties such as `document`, but the actual
value is undefined. This exact situation has been occurring and is
causing the instrumentDOM function to throw an error as `'document' in
WINDOW` is technically true though the value is falsey. We should rather
attempt an actual check of the value being truthy rather than if the
property is defined

Co-authored-by: Luca Forstner <luca.forstner@sentry.io>