feat(core): Add `ServerRuntimeClient` (#8930)
The `ServerRuntimeClient` is a near identical copy of the nextjs
`EdgeClient`. To make it a direct replacement it has constructor options
to override the event `platform`, `runtime`, and `server_name`.

This PR makes yet another copy of the Node `eventbuilder.ts` but after
future PRs to remove the `EdgeClient` and make `NodeClient` extend
`ServerRuntimeClient`, this will be the only copy. I've put the
`eventbuilder` code in utils since some of these functions are used
elsewhere outside of the clients and I don't want to export these from
core and them become part of our public API. This is especially
important since the browser SDK already exports it's own slightly
different `exceptionFromError`.