ref(core): Avoid unnecessary breadcrumbs array mutations (#8957)
This is a micro improvement, but maybe worth it as we can have a lot of
breadcrumbs. This ensures we only create a new breadcrumbs array if we
exceed the limit.

In contrast, currently we'll always create two copies of the breadcrumbs
for each added breadcrumb (first for the array spread, then we copy it
through the slice), even if we don't really need to do this.