feat(core): Update span performance API names (#8971)
As per the new changes in RFC 101 in
https://github.com/getsentry/rfcs/pull/113, update the span performance
API names.

- `startActiveSpan` -> `startSpan`
- `startSpan` -> `startInactiveSpan`

https://github.com/getsentry/rfcs/blob/main/text/0101-revamping-the-sdk-performance-api.md

`startActiveSpan` is deprecated, while `startInactiveSpan` is being
introduced. The breaking change is that `startSpan` is being changed,
but considering that basically no-one is using the `startSpan` API, we
should be fine to break here for correctness reasons. Better break now
than to have everyone refactor their code in v8.