fix(replay): Ensure `handleRecordingEmit` aborts when event is not added (#8938)
I noticed that in `handleRecordingEmit`, due to the async nature of
`addEvent` it could happen that an event is actually not added (because
it is discarded due to timestamp etc.). But we are still updating the
initial session timestamp (`[Replay] Updating session start time to
earliest event in buffer to...`), as we don't actually abort there.

This PR changes this to actually abort `handleRecordingEmit` in this
case. I added an `addEventSync` method for this that just returns
true/false instead of a promise, which should not change anything there
as we haven't been waiting for the result of the promise anyhow.