fix(node-otel): Refactor OTEL span reference cleanup (#9000)
This PR updates the span reference cleanup to take into account if a
span is/may still be referenced somewhere else.

Previously, whenever a span finished we removed the reference from the
map, to clean up and avoid memory leaks.
However, it seems that sometimes spans are ended before a child span is
started (at least the hooks may fire in this order). This leads to the
potential case where a parent that _should_ exist cannot be found, thus
creating a new transaction instead of a span.

With this change, we keep more information in our span map, in order to
clear sub-spans (=not transactions) only when the root span
(=transaction) is finished.