feat(integrations): Introduce `processEvent` hook on `Integration`
This adds a new (optional) `processEvent` hook on the `Integration` interface, which allows to register an event processor **for the current client only**.

This has actually correct semantics in that the processor will only be registered for the client that the integration is added for. This is done by adding a new `addEventProcessor` method on the baseclient, which for now are called after all global & scope event processors.