feat(core): Introduce `processEvent` hook on `Integration` (#9017)
This adds a new (optional) `processEvent` hook on the `Integration`
interface, which allows to register an event processor **for the current
client only**.

This has actually correct semantics in that the processor will only be
registered for the client that the integration is added for. This is
done by adding a new `addEventProcessor` method on the client, which for
now are called after all global & scope event processors.

Previously, all integrations always registered a _global_ event
processor, which is not really necessary. With this, we can be much more
focused & also skip checking for existence of the integration on the
client etc.