fix(sveltekit): Flush in server wrappers before exiting (#9153)
When deploying to Vercel (non-edge), server code is converted into
Lambda functions. These functions have a tendency to freeze/shut down
before the SDK is able to flush out events, causing events to be
dropped. We already handle this in NextJS by manually calling `flush` 
if we detect we're in a serverless/Lambda environment.
This patch applies the (almost) same logic to SvelteKit's server side
instrumentation.