fix(node-experimental): Sample in OTEL Sampler (#9203)
In order for sampling propagation to work, we need to sample in OTEL,
not in Sentry. As the sentry spans are only created later.

Note that this _does not_ fix propagation yet in node-experimental, once
this is merged I still need to adjust the propagator (or rather, fork
it) from opentelemetry-node to work without sentry spans.