feat(astro): Add `sentryAstro` integration (#9218)
Add the `sentryAstro` Astro integration which will be the core
component of the Astro SDK.

The integration takes care of:

* injecting SDK init code into the client and server entry points. 
* emitting source maps and adding the Sentry Vite plugin for source maps upload

The idea is: Basic SDK setup will only require `sentryAstro` to be added
by users. Everything else happens automatically. We also support a more
custom setup but it will require additional config files. The reason is
that we can only inject serialized code. We cannot take e.g. function
callbacks (e.g. `beforeSend`) in the options of `sentryAstro`.

---------

Co-authored-by: Francesco Novy <francesco.novy@sentry.io>