feat(feedback): Bootstrap new Feedback integration (#9139)
Note this is mostly a placeholder for our new Feedback integration. This is not intended to be published (yet) with the rest of the monorepo packages. It currently has limited functionality. There will be follow-ups to allow creation of the UI widget, and eventually will be added to the Sentry namespace.