feat(opentelemetry): Add new `@sentry/opentelemetry` package (#9238)
This extracts the code from `node-experimental` in a more reusable
`@sentry/opentelemetry` package.

This package has no direct (otel) dependencies, but only peer
dependencies. This means it can be used by anybody with an existing OTEL
setup without messing with their dependencies etc.

In turn, `node-experimental` uses this package to implement a concrete
and opinionated implementation of this for node.

This new package should also be usable for non-node OTEL
implementations, and concisously does not include any node specific
stuff.