feat(replay): Allow to treeshake rrweb features (#9274)
This depends on https://github.com/getsentry/rrweb/pull/114 to be merged
first, but allows to configure build time flags to shake out certain
rrweb features that may not be used.

It also adds a size limit entry that shows the total bundle size with
everything that can be shaken out removed, incl. debug stuff. Bundle
size is about ~11kb gzipped less in this scenario, which is not bad.