test(loader): Improve loader tests & update loader (#9245)
This updates the loader based on this PR:
https://github.com/getsentry/sentry/pull/58070
and also actually fixes/improves the loader tests by avoiding importing
`@sentry/browser`. This is not needed and actually obfuscates the
"real"/"normal" behavior (just look at the generated `dist` folder for a
test.

Instead, I added an eslint rule to allow `Sentry` as a global for these
cases, and just access them directly. This also uncovered an incorrect
test, which I adjusted so it works. I also added a new test to ensure
custom config works.