fix(replay): Fix xhr start timestamps (#9341)
We've been using the wrong `startTimestamp` for the core xhr
instrumentation. Outside of replay, this wasn't noticed because we are
not actually using the anywhere 😬 But in replay, it lead to all xhr
breadrcumbs showing an incorrect duration of `0`.

Note that this is _maybe_ not 100% correct, as in theory you could call
`xhr.send()` later, which is probably the _most correct_ start
timestamp. But this would require us to keep the start time somewhere on
the xhr object, which is a bit trickier than this solution. So I think
it is fine to do this based on `xhr.open()` (and _definitely_ more
correct than it was before).

fixes https://github.com/getsentry/sentry/issues/52790